requirements-minimal.txt
attrs
click
matplotlib
opencv-contrib-python-headless
scikit-learn
tensorflow>=2.0.0
